fix bug author null change the author with commiter.
The author is the person who originally wrote the code.
The committer, on the other hand,
is assumed to be the person who committed the code on behalf of the original author.
This is important in Git because Git allows you to rewrite history,
or apply patches on behalf of another person.